#66 combine migration, role move to user
