+++ BookHub.Contracts/REST/Requests/Users/UpdateUserProfileInfoRequest.cs
@@ -1,7 +1,4 @@
-﻿using BookHub.Models.Users;
-
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
 namespace BookHub.Contracts.REST.Requests.Users;
