+++ BookHub.Models.Tests/Users/UserProfileInfo.Tests.cs
@@ -18,17 +18,15 @@ public class UserProfileInfoTests
         var name = new Name<User>("name");
         var email = new MailAddress("email@gmail.com");
         var about = new About("some");
-        var userRole = UserRole.Moderator;
 
         // Act
-        var profileInfo = new UserProfileInfo(id, name, email, about, userRole);
+        var profileInfo = new UserProfileInfo(id, name, email, about);
 
         // Assert
         profileInfo.Id.Should().Be(id);
         profileInfo.Name.Should().Be(name);
         profileInfo.Email.Should().Be(email);
         profileInfo.About.Should().Be(about);
-        profileInfo.Role.Should().Be(userRole);
     }
 
     [Fact(DisplayName = "Cannot create without id.")]
@@ -41,8 +39,7 @@ public class UserProfileInfoTests
                 null!,
                 new("name"),
                 new("email@gmail.com"),
-                new("about"),
-                UserRole.Moderator));
+                new("about")));
 
         // Assert
         exception.Should().BeOfType<ArgumentNullException>();
@@ -58,8 +55,7 @@ public class UserProfileInfoTests
                 new(123),
                 null!,
                 new("email@gmail.com"),
-                new("about"),
-                UserRole.Moderator));
+                new("about")));
 
         // Assert
         exception.Should().BeOfType<ArgumentNullException>();
@@ -75,8 +71,7 @@ public class UserProfileInfoTests
                 new(123),
                 new("name"),
                 null!,
-                new("about"),
-                UserRole.Moderator));
+                new("about")));
 
         // Assert
         exception.Should().BeOfType<ArgumentNullException>();
@@ -92,27 +87,9 @@ public class UserProfileInfoTests
                 new(123),
                 new("name"),
                 new("email@gmail.com"),
-                null!,
-                UserRole.Moderator));
+                null!));
 
         // Assert
         exception.Should().BeOfType<ArgumentNullException>();
-    }
-
-    [Fact(DisplayName = "Cannot create with invalid user role.")]
-    [Trait("Category", "Unit")]
-    public void CanNotCreateWithoInvalidRole()
-    {
-        // Act
-        var exception = Record.Exception(() =>
-            new UserProfileInfo(
-                new(123),
-                new("name"),
-                new("email@gmail.com"),
-                new("about"),
-                (UserRole)int.MaxValue));
-
-        // Assert
-        exception.Should().BeOfType<InvalidEnumArgumentException>();
     }
 }
