+++ BookHub.Models/Users/User.cs
@@ -10,10 +10,13 @@ public sealed class User
 {
     public UserProfileInfo ProfileInfo { get; }
 
+    public UserRole Role { get; }
+
     public UserStatus Status { get; }
 
     public User(
         UserProfileInfo profileInfo,
+        UserRole role,
         UserStatus status)
     {
         ArgumentNullException.ThrowIfNull(profileInfo);
@@ -27,5 +30,14 @@ public sealed class User
                 typeof(UserStatus));
         }
         Status = status;
+
+        if (!Enum.IsDefined(role))
+        {
+            throw new InvalidEnumArgumentException(
+                nameof(role),
+                (int)role,
+                typeof(UserRole));
+        }
+        Role = role;
     }
 }
