+++ BookHub.Models/Users/UserProfileInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.ComponentModel;
 using System.Net.Mail;
 
 namespace BookHub.Models.Users;
@@ -17,14 +16,11 @@ public sealed class UserProfileInfo
 
     public About About { get; }
 
-    public UserRole Role { get; }
-
     public UserProfileInfo(
         Id<User> id,
         Name<User> name,
         MailAddress email,
-        About about,
-        UserRole role)
+        About about)
     {
         ArgumentNullException.ThrowIfNull(id);
         Id = id;
@@ -37,15 +33,5 @@ public sealed class UserProfileInfo
 
         ArgumentNullException.ThrowIfNull(about);
         About = about;
-
-        if (!Enum.IsDefined(role))
-        {
-            throw new InvalidEnumArgumentException(
-                nameof(role),
-                (int)role,
-                typeof(UserRole));
-        }
-        Role = role;
-
     }
 }
